typescript linting errors cleared
